Create README.md
The small project allows a minimote device to control 2 separate dimmers.

The buttons on the left control one dimmer the ones on the right control another dimmer.

The UP button increase the light level of 20%
The DOWN button lower light level of 20%

If the up or down buttons are hold, the light is turned off.

The light is set to a specific level when the light is turned on.  This level is set when the application is setup.